pyproject.toml
[project]
name = "autofpl"
version = "0.1.0"
description = "Automated FPL pipeline: fetch data, score players, LLM decisions, optional apply."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31",
    "google-generativeai>=0.8",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.scripts]
autofpl = "autofpl.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autofpl"]
